--- a/README.md
+++ b/README.md
@@ -1,5 +1,3 @@
-[![CircleCI](https://circleci.com/gh/nikkolasg/snarkpack/tree/main.svg?style=svg)](https://circleci.com/gh/nikkolasg/snarkpack/tree/main)
-
 # Snarpack on arkwork
 
 This is a port in the arkwork framework of the original [implementation in bellperson](https://github.com/filecoin-project/bellperson/tree/master/src/groth16/aggregate) of [Snarkpack](https://eprint.iacr.org/2021/529.pdf). Note both works are derived from the original arkwork implementation of the inner pairing product argument (IPP) [paper](https://eprint.iacr.org/2019/1177.pdf).