--- a/benches/aggregation/mod.rs
+++ b/benches/aggregation/mod.rs
@@ -1,8 +1,6 @@
 use ark_bls12_381::{Bls12_381, Fr};
 use ark_ff::One;
-use ark_groth16::{
-    create_random_proof, generate_random_parameters, prepare_verifying_key, verify_proof,
-};
+use ark_groth16::{prepare_verifying_key, Groth16};
 use snarkpack;
 use snarkpack::transcript::Transcript;
 
@@ -16,7 +14,7 @@ pub fn groth16_aggregation() {
     let mut rng = rand_chacha::ChaChaRng::seed_from_u64(1u64);
     let params = {
         let c = Benchmark::<Fr>::new(num_constraints);
-        generate_random_parameters::<Bls12_381, _, _>(c, &mut rng).unwrap()
+        Groth16::<Bls12_381>::generate_random_parameters_with_reduction(c, &mut rng).unwrap()
     };
     // prepare the verification key
     let pvk = prepare_verifying_key(&params.vk);
@@ -28,13 +26,14 @@ pub fn groth16_aggregation() {
     let proofs = (0..nproofs)
         .map(|_| {
             let c = Benchmark::new(num_constraints);
-            create_random_proof(c, &params, &mut rng).expect("proof creation failed")
+            Groth16::<Bls12_381>::create_random_proof_with_reduction(c, &params, &mut rng)
+                .expect("proof creation failed")
         })
         .collect::<Vec<_>>();
     // verify we can at least verify one
     let inputs: Vec<_> = [Fr::one(); 2].to_vec();
     let all_inputs = (0..nproofs).map(|_| inputs.clone()).collect::<Vec<_>>();
-    let r = verify_proof(&pvk, &proofs[1], &inputs).unwrap();
+    let r = Groth16::<Bls12_381>::verify_proof(&pvk, &proofs[1], &inputs).unwrap();
     assert!(r);
 
     let mut prover_transcript = snarkpack::transcript::new_merlin_transcript(b"test aggregation");