Test benchmarking bot