@achimcc 2023-06-07
## Benchmark for 8f34519
  <details>
    <summary>Click to view benchmark</summary>

| Test | Base         | PR               | % |
|------|--------------|------------------|---|
| groth16 aggregation | 1614.9±23.23ms | 1614.0±24.20ms | -0.06% |

  </details>